Synthesize and perform depth-sweep on all xilinx bugs
To make them synthesizable in vivado, each bug is equipped with
1. A tcl script to mimic the process of creating, synthesizing and
report utilizations in the vivado
2. A Verilog wrapper to contain the systemverilog code generated from
our tools. Note that except AXI-Lite-DropAck and
AXI-Stream-Incorrect-Last, the wrappers of all other bugs have a
AXI-Lite interface "unnaturally" wired to each application. Those
wrappers do not comply AXI-Lite protocol and only created for synthesis.

The list of Xilinx bugs included in this commit:
- AXI-Lite-DropAck
- AXI-Stream-Incorrect-Last
- SDSPI-path-merging
- SDSPI-deadlock
- SDSPI-Endian
- FFT-BitTruncation
- FADD-Misindexing